requirements.txt
attrdict==2.0.1
bcrypt==3.1.7
Cerberus==1.3.2
cffi==1.14.0
cmake==3.17.2
colorzero==1.1
cryptography==3.2
dlib==19.16.0
gpiozero==1.5.1
localconfig==1.1.3
numpy==1.14.2
opencv-python==4.2.0.34
paramiko==2.7.1
pycparser==2.20
PyNaCl==1.4.0
pysftp==0.2.9
PyYAML==5.4
requests==2.26.0
six==1.15.0
